New Distances and organization
- Addition of Taxicab and Chebyshev Distance Functions
- Moved some Datatype related functions to datatype files
