+++ src/misc/version.h
@@ -4,9 +4,9 @@
 #define PROJECT_VERSION_MINOR 2
 #define PROJECT_VERSION_PATCH 7
 #define PROJECT_VERSION_STRING "0.2.7"
-#define PROJECT_VERSION_FULL_STRING "0.2.7 (latest/1afa80d)"
+#define PROJECT_VERSION_FULL_STRING "0.2.7 (latest/5b7ef1f)"
 #define PROJECT_NAME "BetrockServer"
-#define PROJECT_GIT_COMMIT "1afa80d"
+#define PROJECT_GIT_COMMIT "5b7ef1f"
 #define PROJECT_GIT_BRANCH "latest"
 #define PROJECT_NAME_VERSION "BetrockServer 0.2.7"
-#define PROJECT_NAME_VERSION_FULL "BetrockServer 0.2.7 (latest/1afa80d)"
+#define PROJECT_NAME_VERSION_FULL "BetrockServer 0.2.7 (latest/5b7ef1f)"
